requirements.txt
azure-keyvault-secrets 
azure-identity